+++ src/raycaster-core.js
@@ -31,7 +31,7 @@ export function Raycaster(options) {
     * @readonly
     * @since 0.6.0
     */
-    this.version = '0.11.0';
+    this.version = '0.10.6';
     /**
     * Raycaster's scene
     *
@@ -139,7 +139,7 @@ export function Raycaster(options) {
     *
     * @name Raycaster#dynamicMappedObjects
     * @type {object[]}
-    * @since 0.11.0
+    * @since 0.10.6
     */
      this.dynamicMappedObjects = [];
     /**
@@ -174,11 +174,11 @@ export function Raycaster(options) {
 
         if(options.autoUpdate === undefined || options.autoUpdate)
             //automatically update event
-            this.scene.events.on('update', this.update.bind(this));
+            this.scene.events.on('update', this.update, this);
     }
     else
         //automatically update event
-        this.scene.events.on('update', this.update.bind(this));
+        this.scene.events.on('update', this.update, this);
 
     return this;
 }
@@ -594,14 +594,14 @@ Raycaster.prototype = {
 
         if(!this.debugOptions.maps)
             return this;
-
+            
         for(let object of this.mappedObjects)
         {
             let map;
         
             if(object.type === 'body' || object.type === 'composite')
                 map = object.raycasterMap;
-            else
+            else if(object.data)
                 map = object.data.get('raycasterMap');
             
             if(!map)
@@ -643,6 +643,13 @@ Raycaster.prototype = {
      */
     destroy: function() {
         this.removeMappedObjects(this.mappedObjects);
+        
+        if(this.graphics)
+            this.graphics.destroy();
+        
+        if(this.scene) {
+            this.scene.events.removeListener('update', null, this);
+        }
 
         for(let key in this) {
             delete this[key];